Do not cull things we need for this run of ybd

+++ ybd/cache.py
@@ -74,6 +74,7 @@ def cache_key(defs, this):
             app.log('cache-log', 'cache logged to',
                     app.config.get('cache-log'))
 
+    app.config['keys'] += [definition['cache']]
     return definition['cache']
 
 
@@ -300,7 +301,7 @@ def cull(artifact_dir):
             path = os.path.join(artifact_dir, artifact)
             if os.path.exists(os.path.join(path, artifact + '.unpacked')):
                 path = os.path.join(path, artifact + '.unpacked')
-            if os.path.exists(path):
+            if os.path.exists(path) and not artifact in app.config['keys']:
                 tmpdir = tempfile.mkdtemp()
                 shutil.move(path, os.path.join(tmpdir, 'to-delete'))
                 app.remove_dir(tmpdir)
